refactor - rm comentarios/logs
