+++ front-end/src/pages/CustomerCheckout/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function CheckoutComprador() {
+  return (
+    <div>
+      <h1>CheckoutComprador</h1>
+    </div>
+  );
+}
